pyproject.toml
[tool.poetry]
name = "matthew-bot"
version = "0.1.0"
description = "a bot that replicates messages from my contact matthew"
authors = ["Amar Paul <amar.paul16@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.10"
pytz = "^2021.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
